+++ src/com/company/pixel/BufferedImagePixelDrawer.java
@@ -3,7 +3,7 @@ package com.company.pixel;
 import java.awt.*;
 import java.awt.image.BufferedImage;
 
-public class BufferedImagePixelDrawer implements PixelDrawer{
+public class BufferedImagePixelDrawer implements PixelDrawer {
     private BufferedImage bi;
 
     public BufferedImagePixelDrawer(BufferedImage bi) {
@@ -11,10 +11,10 @@ public class BufferedImagePixelDrawer implements PixelDrawer{
     }
 
     @Override
-    public void drawPixel(int x, int y, Color c) {
+    public void drawPixel(double x, double y, Color c) {
 
         if (x >= 0 && y >= 0 && x < bi.getWidth() && y < bi.getHeight()) {
-            bi.setRGB(x, y, c.getRGB());
+            bi.setRGB((int) x, (int) y, c.getRGB());
         }
 
     }
