+++ src/com/company/pixel/PixelDrawer.java
@@ -3,6 +3,6 @@ package com.company.pixel;
 import java.awt.*;
 
 public interface PixelDrawer {
-    void drawPixel(int x, int y, Color c);
+    void drawPixel(double x, double y, Color c);
 
 }
